Refactor helpers into grouped modules
