+++ src/helpers/carouselBuilder.js
@@ -2,10 +2,9 @@ import { createGokyoLookup } from "./utils.js";
 import { generateJudokaCard } from "./cardBuilder.js";
 import { getFallbackJudoka } from "./judokaUtils.js";
 import { setupLazyPortraits } from "./lazyPortrait.js";
-import { createScrollButton, updateScrollButtonState } from "./carouselScroll.js";
+import { createScrollButton, updateScrollButtonState, setupKeyboardNavigation, setupSwipeNavigation, applyAccessibilityImprovements } from "./carousel/index.js";
 import {
   CAROUSEL_SCROLL_DISTANCE,
-  CAROUSEL_SWIPE_THRESHOLD,
   SPINNER_DELAY_MS
 } from "./constants.js";
 import { getMissingJudokaFields, hasRequiredJudokaFields } from "./judokaValidation.js";
@@ -130,96 +129,6 @@ function handleBrokenImages(card) {
       img.src = "./assets/cardBacks/cardBack-2.png";
     };
   }
-}
-
-/**
- * Sets up keyboard navigation for the carousel container.
- *
- * @pseudocode
- * 1. Make the container focusable by setting `tabIndex` to 0.
- * 2. Add a `keydown` event listener to the container.
- *    - Scroll left when the "ArrowLeft" key is pressed and focus the previous card.
- *    - Scroll right when the "ArrowRight" key is pressed and focus the next card.
- *
- * @param {HTMLElement} container - The carousel container element.
- */
-function setupKeyboardNavigation(container) {
-  container.tabIndex = 0; // Make the carousel focusable
-  const cards = container.querySelectorAll(".judoka-card"); // Cache the NodeList
-  container.addEventListener("keydown", (event) => {
-    const active = document.activeElement;
-    const index = Array.from(cards).indexOf(active);
-
-    if (event.key === "ArrowLeft") {
-      container.scrollBy({ left: -CAROUSEL_SCROLL_DISTANCE, behavior: "smooth" });
-      const prevIndex = index > 0 ? index - 1 : 0;
-      cards[prevIndex]?.focus();
-    } else if (event.key === "ArrowRight") {
-      container.scrollBy({ left: CAROUSEL_SCROLL_DISTANCE, behavior: "smooth" });
-      const nextIndex = index >= 0 ? Math.min(cards.length - 1, index + 1) : 0;
-      cards[nextIndex]?.focus();
-    }
-  });
-}
-
-/**
- * Sets up swipe navigation for the carousel container.
- *
- * @pseudocode
- * 1. Track the starting X position of a touch event (`touchstart`).
- * 2. Track the ending X position of a touch event (`touchend`).
- * 3. Calculate the swipe distance and direction.
- *    - Scroll left if the swipe distance is greater than 50 pixels.
- *    - Scroll right if the swipe distance is less than -50 pixels.
- *
- * @param {HTMLElement} container - The carousel container element.
- */
-function setupSwipeNavigation(container) {
-  let touchStartX = 0;
-
-  container.addEventListener("touchstart", (event) => {
-    touchStartX = event.touches[0].clientX;
-  });
-
-  container.addEventListener("touchend", (event) => {
-    const touchEndX = event.changedTouches[0].clientX;
-    const swipeDistance = touchEndX - touchStartX;
-
-    if (swipeDistance > CAROUSEL_SWIPE_THRESHOLD) {
-      container.scrollBy({ left: -CAROUSEL_SCROLL_DISTANCE, behavior: "smooth" });
-    } else if (swipeDistance < -CAROUSEL_SWIPE_THRESHOLD) {
-      container.scrollBy({ left: CAROUSEL_SCROLL_DISTANCE, behavior: "smooth" });
-    }
-  });
-}
-
-// Ensure WCAG compliance for touch target sizes and contrast ratios
-
-// Adjust button sizes for touch targets
-const MIN_TOUCH_TARGET_SIZE = 48;
-
-function ensureTouchTargetSize(element) {
-  const style = window.getComputedStyle(element);
-  const width = parseInt(style.width, 10);
-  const height = parseInt(style.height, 10);
-
-  if (width < MIN_TOUCH_TARGET_SIZE || height < MIN_TOUCH_TARGET_SIZE) {
-    element.style.minWidth = `${MIN_TOUCH_TARGET_SIZE}px`;
-    element.style.minHeight = `${MIN_TOUCH_TARGET_SIZE}px`;
-    element.style.padding = "10px"; // Add padding for better usability
-  }
-}
-
-// Apply touch target size adjustments to scroll buttons
-function applyAccessibilityImprovements(wrapper) {
-  const buttons = wrapper.querySelectorAll(".scroll-button");
-  buttons.forEach((button) => ensureTouchTargetSize(button));
-
-  // Ensure contrast ratios for text and background
-  const cards = wrapper.querySelectorAll(".judoka-card");
-  cards.forEach((card) => {
-    card.style.color = "var(--color-text-inverted)";
-  });
 }
 
 /**
