+++ src/helpers/classicBattle.js
@@ -14,6 +14,7 @@ import {
 } from "./battleEngine.js";
 import * as infoBar from "./setupBattleInfoBar.js";
 
+import { getStatValue, updateScoreDisplay } from "./battle/index.js";
 function getCountdown() {
   if (typeof window !== "undefined" && window.startCountdownOverride) {
     return window.startCountdownOverride;
@@ -58,20 +59,6 @@ function resetStatButtons() {
   });
 }
 
-function getStatValue(container, stat) {
-  const index = STATS.indexOf(stat) + 1;
-  const span = container.querySelector(`li.stat:nth-child(${index}) span`);
-  return span ? parseInt(span.textContent, 10) : 0;
-}
-
-function updateScoreDisplay() {
-  const { playerScore, computerScore } = getScores();
-  infoBar.updateScore(playerScore, computerScore);
-  const el = document.getElementById("score-display");
-  if (el) {
-    el.innerHTML = `You: ${playerScore}<br>\nComputer: ${computerScore}`;
-  }
-}
 
 function showResult(message) {
   const el = document.getElementById("round-message");
