+++ tests/card/cardBuilder.test.js
@@ -1,4 +1,4 @@
-import { createScrollButton } from "../../src/helpers/carouselScroll.js";
+import { createScrollButton } from "../../src/helpers/carousel/scroll.js";
 import { generateCardPortrait, generateCardStats } from "../../src/helpers/cardRender.js";
 import { vi } from "vitest";
