+++ tests/helpers/countrySlider.test.js
@@ -20,7 +20,7 @@ describe("createCountrySlider", () => {
       c.appendChild(slideB);
     });
 
-    vi.doMock("../../src/helpers/countryUtils.js", () => ({ populateCountryList }));
+    vi.doMock("../../src/helpers/country/index.js", () => ({ populateCountryList }));
 
     const { createCountrySlider } = await import("../../src/helpers/countrySlider.js");
